combine tests for rule lowering
