+++ frontend/test/pytest/test_rule_lowering.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Tests for the decomposition rule precompilation utilities."""
+"""Unit tests for the python decompositions module."""
 
 from pathlib import Path
 
@@ -20,6 +20,7 @@ import pennylane as qp
 import pytest
 
 from catalyst.compiler import _quantum_opt
+from catalyst.device.python_decompositions import python_decomposition_wrapper
 from catalyst.utils.precompile_decomposition_rules import (
     compile_op_decomp_rules,
     get_abstract_args,
@@ -28,9 +29,7 @@ from catalyst.utils.precompile_decomposition_rules import (
 from catalyst.utils.runtime_environment import BYTECODE_FILE_PATH
 
 
-class TestGetAbstractArgs:
-    """Tests for get_abstract_args."""
-
+class TestPrecompiled:
     def test_ignore_wires(self):
         """Test that get_abstract_args correctly ignores WiresLike params."""
         assert not get_abstract_args(qp.X)
@@ -55,8 +54,6 @@ class TestGetAbstractArgs:
         with pytest.raises(ValueError, match="Cannot generate arguments"):
             get_abstract_args(qp.ControlledQubitUnitary)
 
-
-class TestCompileOpDecompRules:
     """Tests for compile_op_decomp_rules."""
 
     def test_compile_hadamard_rules(self):
@@ -126,33 +123,73 @@ class TestCompileOpDecompRules:
 
                 compile_op_decomp_rules(NewFakeOp)
 
+    def test_bytecode_file(self):
+        """Test that the bytecode file is generated correctly."""
+        orig_bcfile = Path(BYTECODE_FILE_PATH)
+        tmp_bcfile = None
 
-def test_bytecode_file():
-    """Test that the bytecode file is generated correctly."""
-    orig_bcfile = Path(BYTECODE_FILE_PATH)
-    tmp_bcfile = None
+        if orig_bcfile.exists():
+            tmp_bcfile = orig_bcfile.replace(BYTECODE_FILE_PATH + ".tmpbackup")
 
-    if orig_bcfile.exists():
-        tmp_bcfile = orig_bcfile.replace(BYTECODE_FILE_PATH + ".tmpbackup")
+        try:
+            precompile_decomp_rules()
+            assert orig_bcfile.exists()
 
-    try:
-        precompile_decomp_rules()
-        assert orig_bcfile.exists()
+        finally:
+            if tmp_bcfile:
+                tmp_bcfile = tmp_bcfile.replace(orig_bcfile)
+            else:
+                orig_bcfile.unlink(missing_ok=True)
 
-    finally:
-        if tmp_bcfile:
-            tmp_bcfile = tmp_bcfile.replace(orig_bcfile)
-        else:
-            orig_bcfile.unlink(missing_ok=True)
+        # NOTE: empty pass is needed to prevent running default pipeline
+        rules = _quantum_opt("--empty", BYTECODE_FILE_PATH)
 
-    # NOTE: empty pass is needed to prevent running default pipeline
-    rules = _quantum_opt("--empty", BYTECODE_FILE_PATH)
+        assert "_isingxy_to_h_cy" in rules
+        assert "_doublexcit" in rules
+        assert "_pauliz_to_ps" in rules
+        assert "_cphase_to_ppr" in rules
+        assert "_crot" in rules
 
-    assert "_isingxy_to_h_cy" in rules
-    assert "_doublexcit" in rules
-    assert "_pauliz_to_ps" in rules
-    assert "_cphase_to_ppr" in rules
-    assert "_crot" in rules
+
+class TestTraceTime:
+    pass
+
+
+class TestOnDemand:
+    """
+    Test the python wrapper functions used for on-demand, compile-time decomposition rule lowering.
+    """
+
+    def test_paulirot_wrapper(self):
+        """Test that the paulirot QPD wrapper correctly returns the IR as a string."""
+        result = python_decomposition_wrapper(
+            "PauliRot", "PauliRot[f64][3]{pauli_word:XZZ}", ["i32"], [3], {"pauli_word": "XZZ"}
+        )
+        assert isinstance(result, str)
+        assert "_pauli_rot_decomposition" in result
+        assert 'target_gate = "PauliRot[f64][3]{pauli_word:XZZ}"' in result
+        assert "Hadamard" in result
+        assert "multirz" in result
+
+    def test_multiple_rules(self):
+        """Test that the python decomposition wrapper supports multiple rules."""
+        with qp.decomposition.local_decomps():
+
+            def test_resources(pauli_word):
+                return {qp.X: 1}
+
+            @qp.register_resources(test_resources)
+            def test_decomp(angle, wires, pauli_word):
+                qp.RX(angle, wires[0])
+
+            qp.add_decomps(qp.PauliRot, test_decomp)
+
+            result = python_decomposition_wrapper(
+                "PauliRot", "PauliRot[f64][3]{pauli_word:XYX}", ["f64"], [3], {"pauli_word": "XYX"}
+            )
+
+            assert "test_decomp" in result
+            assert 'target_gate = "PauliRot[f64][3]{pauli_word:XYX}"'
 
 
 if __name__ == "__main__":
